refactor: DTO를 사용하는 컨트롤러 하위로 이동
