+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -3,10 +3,10 @@ package roomescape.theme.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import roomescape.theme.controller.dto.PopularThemeResponse;
+import roomescape.theme.controller.dto.ThemeAvailableTimeResponse;
+import roomescape.theme.controller.dto.ThemeResponse;
 import roomescape.theme.domain.Theme;
-import roomescape.theme.dto.PopularThemeResponse;
-import roomescape.theme.dto.ThemeAvailableTimeResponse;
-import roomescape.theme.dto.ThemeResponse;
 import roomescape.theme.service.AvailableTime;
 import roomescape.theme.service.PopularTheme;
 import roomescape.theme.service.ThemeService;
