+++ src/main/java/roomescape/theme/controller/dto/PopularThemeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.theme.dto;
+package roomescape.theme.controller.dto;
 
 import roomescape.theme.service.PopularTheme;
